Add DSL for registering checks for vitals checks
I did this because we need the consuming application to have the ability to
define how to check its dependencies. This does include dependencies that are
considered critical and required to be checked in the /health request. Ideally
you won't be adding any critical checks, but if you need to, this adds that
ability. The Check Registrar holds the collection of check objects with the
defined name and block to be ran when evaluating the check.