Add the ability to create critical checks for /health
I did this because many legacy applications will not be able to handle
requests correctly if their dependencies are "down". This allows the
middleware to have the flexibility to process registered checks in the process
of making a request to /health.

This includes the ability to set 3 states for a check, `up_state`,
`warn_state`, and `down_state`.