Evaluator default state is down with method name changes
I did this do create the evaluator with a default state of down. This will
make it so that if something goes wrong with the gem based on a funky check
that is created in the application, that the endpoints will report down. For
example if they define a check that does not call `up`, `down`, or `warn` then
the state of the check will default to down since you made an incorrect check

I also switched the methods to `up`, `down`, and `warn` instead of including
the word `state` because even though I thought it was adding more meaning, I
think it was unnecessary.