Registrar is instanced, Evaluator is singleton
I did this because we want to use the Rack::Vitals class to hold the state of
the "singleton" of the registrar, and have the evaluator work on check and
return check result objects.

The registrar is memoized in the Vitals class so that it can be returned, and
passed around instead of being made global.

The CheckEvaluator needs to take checks, run the evaluation and build and
return a check result. The result holds the state of the check and can be
asked about that run of the check. If you want to run it again, you run a
check through the evaluator and it will produce a new result object.